_dispatch_muxnote_release should be an atomic release
Currently, decrementing the reference count of a muxnote is an atomic
operation with relaxed memory order, leading to a race conditions.

`_dispatch_event_loop_drain` dequeues a completion event. Before a
completion event is posted, the reference count of a muxnote is
incremented with `_dispatch_muxnote_retain`. This is an atomic increment
with relaxed memory order.

Once received, `_dispatch_event_loop_drain` may call
`_dispatch_event_merge_pipe_handle_read` which releases the muxnote. The
comment states that the muxnote is "[r]etained when posting the
completion packet", but this may happen on another thread. Therefore,
when `_dispatch_muxnote_release` is called, the retain operation may not
be visible to the thread performing the release, resulting in a disposal
of the muxnote despite still being used.